recommender_api/requirements.txt
apispec==5.1.1
apispec-webframeworks==0.5.2
attrs==21.4.0
boto3==1.21.2
botocore==1.24.2
certifi==2022.12.7
cffi==1.15.0
click==8.0.3
cryptography==38.0.3
fasttext==0.9.2
filelock==3.6.0
Flask==2.0.3
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
huggingface-hub==0.11.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.2.0
jsonschema==4.4.0
lz4==4.0.0
MarkupSafe==2.1.0
marshmallow==3.14.1
more-itertools==8.12.0
moto==3.0.4
numpy==1.22.2
openapi-schema-validator==0.2.3
openapi-spec-validator==0.4.0
packaging==21.3
pandas==1.4.1
platformdirs==2.5.0
psycopg2-binary==2.9.3
pybind11==2.9.1
pycparser==2.21
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2021.3
PyYAML==6.0
rank-bm25==0.2.2
regex==2022.1.18
requests==2.27.1
responses==0.18.0
s3transfer==0.5.1
sacremoses==0.0.47
scikit-learn==1.1.1
six==1.16.0
tokenizers==0.11.5
toml==0.10.2
tomli==2.0.1
torch==1.13.1
tqdm==4.62.3
transformers==4.24.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.1.1
wrapt==1.13.3
xgboost==1.7.1
xlrd==2.0.1
xmltodict==0.12.0
zope.event==4.5.0
zope.interface==5.4.0
